Add phone and text area
Add a phone with a music app and time, it also has a working clock and home button. Area for the dialogue. Its logic wont work until the actual dialogue system is up and running. Add managers like one for UI and one for the phone and one for the text area. Add scriptable object for the songs, they contain the name of the song the author/writer of the song, audio clip/track, its duration and an Icon.